Executor: Fix forceful cancellation
[why]
When a Sequence runs asynchronously and produces so much messages that
the message queue gets clogged up we can not force-stop the Sequence or
destruct the Executor anymore.

A)
The reason is that we can not call cancel() on the Executor. The
executing thread is already waiting for a slot in the queue. But when we
call cancel() the Executor just waits for the thread to become ready
(joinable). That will never happen, because it waits in the queue, ...

B)
When we want to cancel the Sequence this is detected too late (or
never) under certain circumstances, because it is only checked in the
running Lua VM debug hook.

[how]
A)
Distinguish between 'last resort' cancel() and 'orderly' cancel().
Introduce a new function for the later, that takes a Sequence reference
and first empties the message queue into that Sequence, and thus allows
the executing thread to drop all the pending messages until it notices
that it shall stop, and then stop and join.

The ordinary cancel() now just drops any (most) pending messages to
enable the executing thread to finish. That is needed for situations
where we do not know which Sequence is run (i.e. in the destructor).
It might be useful for users, otherwise it could be private.

B)
Check if the Sequence shall terminate before executing the next pending
Step.

Fixes: #54

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>