Replace "LUA" by "Lua"
[why]
To cite from https://www.lua.org/about.html:
"Lua" (pronounced LOO-ah) means "Moon" in Portuguese. As such,
it is neither an acronym nor an abbreviation, but a noun. More
specifically, "Lua" is a name, the name of the Earth's moon
and the name of the language. Like most names, it should be
written in lower case with an initial capital, that is, "Lua".
Please do not write it as "LUA", which is both ugly and
confusing, because then it becomes an acronym with different
meanings for different people. So, please, write "Lua" right!

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>