Executor: Add run_single_step_asynchronously()
[why]
This is a simple way for users to implement single-step
execution within an existing sequence, which is often
required for debugging steps.

[how]
Add a Sequence member function

void run_single_step_asynchronously(Sequence& sequence,
    Context context, StepIndex step_index);

to launch single-step execution. Under the hood, both this
function and the old run_sequence_asynchronously() call

void launch_async(
    Sequence& sequence, Context context, OptionalStepIndex step_index);

which in turn takes care of launching the sequence (if step_index == nullopt)
or a single step (if step_index.has_value()) in a separate thread
and passing the necessary data to it.

Co-authored-by: Ulf Fini Jastrow <ulf.fini.jastrow@desy.de>
Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>