Rework context callback architecture
This commit changes the callbacks provided by the Context:
Where there used to be callbacks for producing various types
of output, there is now just one hook (the
default_message_callback) that should be called whenever a
message is being processed. Consequently, this hook only
receives a reference to the processed message as a parameter.

This modification also has an impact on how print() is
implemented: It basically just does the usual work of
stringifying and concatenating its arguments, but then
just proceeds to send a message of type Type::output.
It is up to the user to install a hook that actually
produces output in the desired format (default is "send to
stdout").

Note that this commit does not implement the actual calling
of the callback in send_message() yet as this requires a
signature change. See next commit for that.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>