Make send_message() call callback from Context
[why]
send_message() is now the main hub for calling user hooks.
Its signature needs to change because it needs access to
the Context for that, with a lot of consequential downstream
changes.

Because send_message() is not really tightly coupled with
CommChannel, I have moved it into its own set of files.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>