Add Sequence to Step::execute to check for timeout.
Since we have a global instance of sequence timeout at the Lua level
there is a problem on parallel sequence executed consecutively. The last
sequence that runs wins. A solution is to inject the sequence timeout
and therefore the Sequence itself to the Step that forwards it to Lua
for checking of timeouts in a hook functionality.