SequenceManager: Use relative paths for sequences
[why]
Before, we did not specify if paths returned by list_sequences() were
relative to the current working directory or to the root directory of
the manager. The same is true for load_sequence().

[how]
Make both list_sequences() and load_sequences() use paths that are
relative to the root directory of the manager.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>