Rewrite legacy sequences to disk when renaming them
[why]
Legacy sequences carry their *label* in the folder name and just there.
After we have successfully renamed them to the new schema, that label is
not stored anywhere anymore, so we are in danger of losing it.

[how]
Load the sequence from the renamed folder and store it again with the
old sequence label.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>