Disallow sequence names starting with a period
[why]
Sequence names are designed to be machine-friendly, and as such are
often used as folder or file names. On unixoid operating systems,
file names starting with a period designate hidden files. We might
want to avoid that by requiring that the first character of a
sequence name must not be a period (".").

Closes #98.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>