Refactor error index correction
[why]
The correct_error_index_*() functions were largely redundant and, in the
case of erase(Iter, Iter), not even extracted into a function.

[how]
Create a single member function correct_error_index(Fct) that receives a
std::function. This std::function reveices the error index (if any) and
returns the corrected value.

Proposed-by: Marcus Walla <marcus.walla@desy.de>
Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>