SequenceManager: Simplify file manipulation
[why]
Sometimes we use the throwing form of file manipulation, sometimes the
error code version. It does not really make sense to catch and
immediately throw again, so unify to the error code variant.

Also sometimes the existence of a file is checked before it is removed,
and sometimes it is not checked.
The std::remove*() function checks the presence of the target anyhow and
returns false/zero if there is none. So rely on that, it will check
(again) anyhow. (They do not throw/set error if the file is missing.)

[how]
Throughout the whole file:
- Do not check for existence before
- Use the error-code version everywhere

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>