SequenceManager: Commit creation of (empty) Sequence
[why]
When a Sequence is created (i.e. a folder is created locally) this is
not reflected on the git repo until we actually save it. That might be
much later.
Each local filesystem change should result in a commit.

[how]
After creating an empty sequence store it normally (with commit).
Have a distinct message for the creation to tell it apart from normal
changes/stores.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>