SequenceManager: Rework committing scheme and messages
[why]
We have very specific code to add files to the index or to stage files
for removal that handles it file by file through the complete changeset.
This could fail if we have files in unexpected directories that change,
because the path is not checked. Anyhow human users would never add all
files one by one but add complete subdirectories with all changes.
That approach is also faster for big repositories.

The messages are a bit off. So the empty separator line is missing so
that the commit messages are rendered extremely long by the usual git
frontends. There is also much redundant information in the message.

[how]
Use the equivalent of `git add subdir` to add all (added / changed /
removed) files of that directory (== Sequence) to the stage and commit
it with a structured message: A title that mentions the directory name
(machine name and uid) and a very concise summery of the changes in the
commit message body.

The actual changes can be better displayed by git clients anyhow;
libgit2 seems to be unable to detect renames correctly, but that can be
seen with 3rd party git clients.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>