Step: Fix modification time
[why]
While working on the git storage I noticed that sometimes the tests
failed because steps were detected as 'changed' that were not actually
touched (und thus unexpected to change).

The problem turned out to be the modification time that gets changed
*on loading* the object (i.e. desirializing).
To make sure the Steps are in a coherent state, at various points in
time we run a function that enforces our preconditions.

Unfortunately that function does not look into the Step object but just
sets the expected/good values, which should always be the same as the
currently active ones.
But calling the setter resets the modification time, regardless of
actual change or not.

[how]
Do not change the modification time if the 'disabled' property is
written to and there is no real change. Only on real changes of the
'disabled' property the modification time is updated.

[note]
I'm not sure if we should issue (additionally) a warning. Maybe that
function is also used to correct the disabled property when other steps
higher in the hierarchy are disabled, and then we do not want a warning.

Maybe we want to exclude the 'disabled' property from modification time
relevant completely?

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>