Fix libgit2 version problem
[why]
Obviously the globbing behavior changed between libgit2 version 0.28.4
(on Ubuntu 20.04) and version 1.1.0 (on Ubuntu 22.04).

While the older version can not work with a star glob in add() and
update() it is needed in the newer version.

This was kind of a problem because one or the other CI machine failed
and whatever was changed, because they are not compatible.

[how]
Detect the libgit2 version on configure time and use the appropriate
glob variant.

I did not research to the concrete version where it changed, though.