Clear up gitignore
There should only files be ignored that are artifacts from correct build
processes and development. Normally these files should not go there but
into the build directory.
If they are NOT ignored one can very easy remove the unwanted stuff with
some `git clean -df`, which is hard(er) if all the stuff is ignored.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>