Add {fmt} formatter for Timeout
[why]
When we have a operator<< for streams we probably also want something
for the more modern {fmt}.

[how]
Just add it based on the stream operator.

[note]
Also alphabetical sorting of functions.

While debugging for the current PR I also had some other classes that
could not be easily output and one needed to lick it up or call obscure
string() OR to_string() member functions. Not even the same everywhere.
This is a mess.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>