SequenceManager: Unify the repeated task to commit some file
[why]
In multiple functions we commit the current changes in the filesystem to
the git repository. The few steps that are needed to do that are coded
out in every place. Code duplication! ;-)

[how]
Pull out the core functionality and have a new function do that.

Because in one instance we need to add two directories (instead of just
one) we pass a SmallVector with the directories that shall be included
in the commit.
But a vector can not pass string references but only string_views.
Passing string_views spreads then like a disease through to some
sub-functions.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>