SequenceManager: Perform commits in tight control
[why]
There are multiple ways a change and commit can go wrong. At the moment
we can not recover from a partial commit, e.g. filesystem changed and
then something happened. Worst is when we could add files to the stash
but the commit did not go through. The next user will start with the
messed up filesystem.

[how]
Perform a git reset on each commit start before we change any files. In
this way we have a well defined starting point. If any exception happens
we undo the filesystem changes with another git reset (--hard).

I choose a helper function that takes a closure - in that closure is the
code that actually changes the files on the filesystem. In principle we
could also have had two functions commit_prepare() and commit_perform()
but catching all errors would have been more boilerplate code.
With the solution here all the complex stuff is localized in the
function template. As there are only 5 (five) places that perform
commits I believe the readabilty might be impacted for some developers
but is still ok. The non-clousre solution does also look not nice, but
that in 5 locations instead of one.

A small complication is that not all to-be-commited directories are
known when we initiate the commit, so the closure can return one (1)
additional path, that is also appended to the commit message title.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>