SequenceManager: Fix use-after-free bug
[why]
We have a use-after-free that went unnoticed:

==4746==ERROR: AddressSanitizer: heap-use-after-free on address 0x603000401cb0 at pc 0x7fab18ab7dc9 bp 0x7fff5d0ec0e0 sp 0x7fff5d0ec0d0
READ of size 1 at 0x603000401cb0 thread T0
    #0 0x7fab18ab7dc8 in escape_glob ../src/SequenceManager.cc:382
    #1 0x7fab18ab80c6 in task::SequenceManager::stage_files_in_directory[abi:cxx11](gul14::basic_string_view<char, gul14::char_traits<char, std::char_traits<char> > >) ../src/SequenceManager.cc:414
    #2 0x7fab18abac4c in perform_commit<task::SequenceManager::store_sequence(const task::Sequence&)::<lambda()> > ../include/taskolib/SequenceManager.h:279
    #3 0x7fab18ab639d in task::SequenceManager::store_sequence(task::Sequence const&) ../src/SequenceManager.cc:334

The problem is the cunstruction of the commit-action closure. We have a
temporary object on the lambda as input, and that has been freed when
the closure is finally called some lines later.

[how]
Just use an owning data structure (i.e. std::string).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>