SequenceManager: Simplify perform_commit()
[why]
The code can work with any number of directories but in fact we only
ever need one - two in one case (rename).
Passing a vector around with no real purpose is not needed.

[how]
Change the interface to take one directory only (instead of a vector of
directories) and take it as last and optional parameter.

Pass the to-be-added directory from the action closure in all cases.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>