SequenceManager: Throw if commit goes wrong
[why]
At the moment when some error classes happen on sequence manipulation
they are silently ignored. Usually they should never be able to fail,
but they can fail if access rights for the actual files are wrong. Not
having any kind of feedback is maybe not so nice.

[how]
Evaluate the return value from commit() and throw an error if no files
could be added. There is no way any of our manipulations can ever be
successful and still not touch any file. So it must be an error if no
files are in the commit.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>