SequenceManager: Fix storing unmodified sequences
[why]
When an unmodified Sequence is to be stored an error is thrown.
It is hard (but not impossible) to distinguish a real failure from an
attempt to "store" no change.

[how]
Do not throw on unmodified Sequences but instead return a boolean that
indicates if changes did exist.
Git errors are still exceptions.
This makes the error handling much more easy.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>