Make execute_lua_script() return expected instead of variant
[why]
Due to the lack of an "expected" type in C++17, the function was
implemented as returning a variant. Unfortunately, dealing with
variants is somewhat clumsy and does not convey the intent of having
an expected outcome and an error outcome.

[how]
We have introduced gul14::expected in the meantime and use it as the
return type.