Replace get_sequence_info_from_filename() by member function
This commit replaces the free function get_sequence_info_from_filename()
by the static member function SequenceManager::parse_folder_name(). This
allows us to drop the SequenceInfo struct, which was almost the same as
the SequenceOnDisk struct. We also stop supporting sequence folder names
in the legacy format with a human-readable label in it.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>