deserialize: Escape raw strings before using in error messages
[why]
The strings that we use raw in the error messages can contain everything
and will usually be interpreted as UTF8 strings. So the message can be
misleading or confusing because unprintable characters that prevent a
match can not be seen, etc.

[how]
Escape the raw strings, meaning all chars <32 and >127 are escaped.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>